Made sure difficulty rating showed up
Since we got userDiff and authorDiff, I made the logic go with userDiff with a priority but take authorDiff when there are no comments. I am assuming here that userDiff ONLY should exist after a user rates the difficulty after a recipe is made. I think rn the userDiff is used in place of authorDiff so that should be fixed in upload page I think.